compile fixup

+++ src/c47/programming/lblGtoXeq.c
@@ -689,7 +689,7 @@ static void _putLiteral(uint8_t *literalAddress) {
     default: {
         #if !defined(DMCP_BUILD)
         printf("\nERROR: in _putLiteral() %u is not an acceptable parameter for ITM_LITERAL!\n", *(literalAddress - 1));
-        printf("At address ram + %" PRIu64 "\n", (literalAddress - 1) - (uint8_t *)ram);
+        printf("At address ram + %" PRIu32 "\n", (uint32_t)((literalAddress - 1) - (uint8_t *)ram));
       #endif // !DMCP_BUILD
     }
   }
